Update breadcrumb title of isDashboardNavMenu (#142)
https://github.com/opensearch-project/OpenSearch-Dashboards/pull/4453

Modified the breadcrumbs from the dashboard plugin,
to be consistent with the actual name of the application.

We should consider not hard checking the name of the
breadcrumb but really hooking into the navigation
plugin that provides the ability to do what this
logic is doing without being hardcoded.

https://github.com/opensearch-project/dashboards-reporting/issues/118

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>
(cherry picked from commit 577871d1ded7d478622964ea23a5f61ed4b0f2e2)